Update alg.cpp useless casting deleted

+++ src/alg.cpp
@@ -24,7 +24,7 @@ uint64_t fact(uint16_t n) {
 }
 
 double calcItem(double x, uint16_t n) {
-    return (double) pown(x, n) / fact(n);
+    return pown(x, n) / fact(n);
 }
 
 double expn(double x, uint16_t count) {
